fix(kotile): compute padded atlas spans in Long, and guard before any arithmetic
Follow-up to the CodeRabbit review on PR #44 (the outside-diff finding).

GlyphAtlasPadding.pageSpanPx returned Int, so `count * (cellPx + 2)` wraps for a
large enough cell — a 16-column page needs only cellPx > ~134M. A wrapped span is
NEGATIVE, which would slip under the `<= GL_MAX_TEXTURE_SIZE` check in both
sources and let through exactly the oversized page the check exists to reject. It
now returns Long and computes in Long; padded() narrows back for the pixmap it
allocates, which is safe because both callers check the span first (stated on the
function).

Also moved FreeTypeGlyphSource's guard above the tight-atlas arithmetic it was
sitting below, so nothing is computed or allocated before the size is known good.

Unlike the rest of this change the arithmetic is pure, so it is unit-tested
GL-FREE and actually ran here: GlyphAtlasPaddingTest, 3 specs (page span and cell
origins for square and non-square cells, and the Long span staying exact past Int
range where the Int product wraps to 16). Note the overflow spec pins the
signature at COMPILE time — reverting pageSpanPx to Int fails compilation rather
than reddening a test, which was confirmed by reverting it.

Re-ran both macOS mirrors after the guard move: freetypeVerify and tileSheetVerify
ALL PASSED with numbers identical to before it (▄ emptyPeak 0, █ edges 0.986-1.0;
sprite margin 0, wall edges 1.0). ./gradlew ktlintCheck test green.

+++ kotile/library/src/main/kotlin/com/sletmoe/kotile/display/ascii/GlyphAtlasPadding.kt
@@ -38,11 +38,19 @@ internal object GlyphAtlasPadding {
     /** Width of the extruded border around every cell, in texels. See the object doc for why it is 1. */
     const val GUTTER_PX = 1
 
-    /** The page span, in px, holding [count] cells of [cellPx] each with their gutters. */
+    /**
+     * The page span, in px, holding [count] cells of [cellPx] each with their gutters.
+     *
+     * A `Long`, because this is what the `GL_MAX_TEXTURE_SIZE` guards in both sources compare against and
+     * an `Int` product wraps for an absurd cell size (a 16-column page needs only `cellPx > ~134M`) —
+     * a wrapped, negative span would slip *under* the limit and let the oversized page through, which is
+     * the one thing the guard exists to stop. Every span a guard accepts is far inside `Int` range, so
+     * [padded] narrows it back for the pixmap it allocates.
+     */
     fun pageSpanPx(
         cellPx: Int,
         count: Int,
-    ): Int = count * (cellPx + 2 * GUTTER_PX)
+    ): Long = count.toLong() * (cellPx.toLong() + 2L * GUTTER_PX)
 
     /** Where cell [index] starts in a padded page of [cellPx] cells — the origin of its texture region. */
     fun cellOriginPx(
@@ -58,6 +66,9 @@ internal object GlyphAtlasPadding {
      * The page is straight-alpha (white RGB, coverage in alpha, or full-colour tile art), so blending is
      * off throughout: every copy here must reproduce the source texel verbatim, including a transparent
      * one, rather than compositing it over the cleared page.
+     *
+     * Callers must have checked the padded span against `GL_MAX_TEXTURE_SIZE` first (both do, before any
+     * rasterising) — that bounds it well inside `Int` range, which is what makes the narrowing below safe.
      */
     fun padded(
         tight: Pixmap,
@@ -72,8 +83,8 @@ internal object GlyphAtlasPadding {
         }
         val page =
             Pixmap(
-                pageSpanPx(cellWidthPx, columns),
-                pageSpanPx(cellHeightPx, rows),
+                pageSpanPx(cellWidthPx, columns).toInt(),
+                pageSpanPx(cellHeightPx, rows).toInt(),
                 Pixmap.Format.RGBA8888,
             ).apply {
                 blending = Pixmap.Blending.None
